ID115: use SendNotificationOperation for message and call notifications
